chore: create comment events avoiding duplication
