+++ src/back/models/Event.ts
@@ -1,18 +1,20 @@
 import { Model } from 'decentraland-gatsby/dist/entities/Database/model'
 import { SQL, table } from 'decentraland-gatsby/dist/entities/Database/utils'
 
-import { Event } from '../../shared/types/events'
+import { Event, EventType } from '../../shared/types/events'
 
 export default class EventModel extends Model<Event> {
   static tableName = 'events'
   static withTimestamps = false
   static primaryKey = 'id'
 
+  // TODO: do we want to show comment events on the activity ticker?
   static async getLatest(): Promise<Event[]> {
     const query = SQL`
       SELECT *
       FROM ${table(EventModel)}
       WHERE created_at >= NOW() - INTERVAL '7 day'
+      AND event_type != ${EventType.Commented}
       ORDER BY created_at DESC
     `
     const result = await this.namedQuery<Event>('get_latest_events', query)
@@ -26,5 +28,15 @@ export default class EventModel extends Model<Event> {
       WHERE created_at < NOW() - INTERVAL '7 day'
     `
     await this.namedQuery('delete_old_events', query)
+  }
+
+  static async alreadyRegistered(discourseEventId: string) {
+    const query = SQL`
+      SELECT count(*)
+      FROM ${table(EventModel)}
+      WHERE (event_data ->>'discourse_event_id') = ${discourseEventId}
+    `
+    const count = (await this.namedQuery<{ count: string }>('find_discourse_event_id', query))[0].count
+    return Number(count) !== 0
   }
 }
