+++ src/entities/User/utils.ts
@@ -1,7 +1,7 @@
 import { hashMessage, recoverAddress } from 'ethers/lib/utils'
 import capitalize from 'lodash/capitalize'
 
-import { DiscoursePostInTopic } from '../../clients/Discourse'
+import { DiscoursePostInTopic } from '../../clients/DiscourseTypes'
 import { FORUM_URL } from '../../constants'
 import { clientEnv } from '../../utils/clientEnv'
 import { ProposalComment, ProposalCommentsInDiscourse } from '../Proposal/types'
