+++ src/services/DiscourseService.ts
@@ -3,7 +3,8 @@ import { requiredEnv } from 'decentraland-gatsby/dist/utils/env'
 
 import { UpdateService } from '../back/services/update'
 import { VoteService } from '../back/services/vote'
-import { Discourse, DiscourseComment, DiscoursePost, DiscoursePostInTopic } from '../clients/Discourse'
+import { Discourse } from '../clients/Discourse'
+import { DiscourseComment, DiscoursePost, DiscoursePostInTopic } from '../clients/DiscourseTypes'
 import ProposalModel from '../entities/Proposal/model'
 import { ProposalWithOutcome } from '../entities/Proposal/outcome'
 import * as proposalTemplates from '../entities/Proposal/templates'
