+++ src/services/ProposalService.ts
@@ -5,7 +5,7 @@ import RequestError from 'decentraland-gatsby/dist/entities/Route/error'
 import { DiscordService } from '../back/services/discord'
 import { EventsService } from '../back/services/events'
 import { NotificationService } from '../back/services/notification'
-import { DiscoursePost } from '../clients/Discourse'
+import { DiscoursePost } from '../clients/DiscourseTypes'
 import { SnapshotProposalContent } from '../clients/SnapshotTypes'
 import UnpublishedBidModel from '../entities/Bid/model'
 import CoauthorModel from '../entities/Coauthor/model'
